+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] (February 2026)
+### Changed
+- changed *iDeal* logo and name to *iDeal - Wero*
+
+### Added
+- "Booking Mode" settings to *PayPal* and *Credit Card* to allow either authorization or charge calls
+- customer language to api calls
+
 ## [2.0.1] (January 2026)
 ### Fixed
 - issue with prepayment not marked as paid in JTL WAWI